refactor: migrate Kokoro to ISentenceSynthesizer
